+++ material/src/features/popover/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './popoverTriggerOptions/popoverTriggerOptions.interface';